Split actuators and mechanisms, change PID stuff
Change how PID coefficients are set in Motor and PidMotor, not that it
matters given how often they're used (never).
Move everything that does not wrap a simple actuator (arm, hand, wheel)
out of actuators.py to mechanisms.py.
